Move some grid controller code into extension file
